# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend Challenge",
        "description": "The purpose of this application is transfer value between 2 users.",
        "termsOfService": "",
        "contact": {
            "email": "rro.oliveira@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Efetua o login do usu\u00e1rio para obter o token",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail do usu\u00e1rio",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Senha do usu\u00e1rio",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe as informa\u00e7\u00f5es do token"
                    },
                    "401": {
                        "description": "Usu\u00e1rio sem permiss\u00e3o de acesso."
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "summary": "Exibe as informa\u00e7\u00f5es do usu\u00e1rio logado.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe as informa\u00e7\u00f5es do usu\u00e1rio logado."
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Efetua o logout do usu\u00e1rio.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout efetuado com sucesso."
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "summary": "Atualiza o token.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna as informa\u00e7\u00f5es do token."
                    }
                }
            }
        },
        "/api/transaction": {
            "post": {
                "summary": "Efetua a transfer\u00eancia de valores entre 2 usu\u00e1rios.",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Valor a ser transferido (Formato americano 1000.00)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payer",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Pagador)",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "payee",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Benefici\u00e1rio)",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer\u00eancia efetuada com sucesso."
                    },
                    "400": {
                        "description": "\n     *                                              N\u00e3o \u00e9 permitido efetuar a transf\u00earencia para o mesmo usu\u00e1rio.\n     *                                              N\u00e3o existe saldo suficiente para realizar a transfer\u00eancia.\n     *                                              Cadastro n\u00e3o localizado em nosso sistema ou n\u00e3o tem permiss\u00e3o para efetuar transfer\u00eancia.\n     *                                              Cadastro n\u00e3o localizado em nosso sistema.\n     *                                              N\u00e3o existe saldo suficiente para realizar a transfer\u00eancia.\n     *                                              Transfer\u00eancia n\u00e3o autorizada.\n     *                                              Erro ao registrar o processo de transfer\u00eancia dos valores.\n     *                                              Erro ao enviar a notifica\u00e7\u00e3o de transfer\u00eancia."
                    }
                }
            }
        },
        "/api/transaction-queue": {
            "post": {
                "summary": "Efetua a transfer\u00eancia de valores entre 2 usu\u00e1rios utilizando fila de processamento.",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Valor a ser transferido (Formato americano 1000.00)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payer",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Pagador)",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "payee",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Benefici\u00e1rio)",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer\u00eancia efetuada com sucesso."
                    },
                    "400": {
                        "description": "\n     *                                              N\u00e3o \u00e9 permitido efetuar a transf\u00earencia para o mesmo usu\u00e1rio.\n     *                                              N\u00e3o existe saldo suficiente para realizar a transfer\u00eancia.\n     *                                              Cadastro n\u00e3o localizado em nosso sistema ou n\u00e3o tem permiss\u00e3o para efetuar transfer\u00eancia.\n     *                                              Cadastro n\u00e3o localizado em nosso sistema.\n     *                                              N\u00e3o existe saldo suficiente para realizar a transfer\u00eancia.\n     *                                              Transfer\u00eancia n\u00e3o autorizada.\n     *                                              Erro ao registrar o processo de transfer\u00eancia dos valores.\n     *                                              Erro ao enviar a notifica\u00e7\u00e3o de transfer\u00eancia."
                    }
                }
            }
        },
        "/api/transaction-with-authentication": {
            "post": {
                "summary": "Efetua a transfer\u00eancia de valores entre 2 usu\u00e1rios. O pagador ser\u00e1 o usu\u00e1rio logado na aplica\u00e7\u00e3o.",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Valor a ser transferido (Formato americano 1000.00)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payee",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Benefici\u00e1rio)",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sem resposta de retorno."
                    }
                }
            }
        },
        "/api/transaction-queue-with-authentication": {
            "post": {
                "summary": "Efetua a transfer\u00eancia de valores entre 2  utilizando fila de processamento. O pagador ser\u00e1 o usu\u00e1rio logado na aplica\u00e7\u00e3o.",
                "parameters": [
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Valor a ser transferido (Formato americano 1000.00)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payee",
                        "in": "query",
                        "description": "C\u00f3digo do usu\u00e1rio (Benefici\u00e1rio)",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sem resposta de retorno."
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "Lista todos os usu\u00e1rios da aplica\u00e7\u00e3o.",
                "responses": {
                    "200": {
                        "description": "Listagem com todos os usu\u00e1rios da aplica\u00e7\u00e3o."
                    }
                }
            }
        }
    },
    "definitions": {}
}